requirements.txt
Flask==2.3.3
playwright==1.40.0
beautifulsoup4==4.12.2
requests==2.31.0
google-generativeai==0.3.1
python-dotenv==1.0.0
Pillow==10.1.0
Flask-WTF==1.2.1
